+++ lib/naver/shorturl.rb
@@ -1,10 +1,11 @@
 module Naver # :nodoc:
-  # Naver Shortenurl API
-  class Shortenurl < Client
+  # Naver Shorturl API
+  class Shorturl < Client
     class << self
       # 단축URL
       # @param url [String] 단축하고자 하는 원본 URL
-      def url(url:)
+      # @return [ObjectifiedHash] 단축 결과
+      def shorten(url:)
         params = { url: url }
         parse(JSON.parse(connection.post("/v1/util/shorturl", params).body)["result"])
       end